requirements.txt
altgraph==0.17.4
importlib_metadata==8.5.0
packaging==25.0
pefile==2023.2.7
pillow==10.4.0
ply==3.11
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyserial==3.5
pywin32-ctypes==0.2.3
tomli==2.2.1
zipp==3.20.2
